Formatting fixes, moved tree methods to PlantLoader, added custom leaf for ExampleTree
